Wiki overhaul
To give new runners a better overview over the wiki, I moved a lot of stuff around and created an introduction page, where a quick overview is provided.
I also moved the lurch on controller segement to the movement knowledge base, as I felt it is too niche for the main page.
Finally I added a 'first run guide (WIP) to help new players getting started.